refactor(gui-agent): move defaultNormalizeCoords to defaultImpls.ts for better organization
